Changed primary key type to INTEGER for SQLite

+++ accountant_db_sqlite/currencies.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS currencies (
-  id                          BIGINT PRIMARY KEY,
+  id                          INTEGER PRIMARY KEY,
   name                        TEXT NOT NULL,
   code                        TEXT NOT NULL,
   num                         TEXT,
@@ -9,50 +9,50 @@ CREATE TABLE IF NOT EXISTS currencies (
 );
 
 CREATE TABLE IF NOT EXISTS sources (
-  id                          BIGINT PRIMARY KEY,
+  id                          INTEGER PRIMARY KEY,
   name                        TEXT NOT NULL,
   details                     TEXT
 );
 
 CREATE TABLE IF NOT EXISTS counterparties (
-  id                          BIGINT PRIMARY KEY,
+  id                          INTEGER PRIMARY KEY,
   name                        TEXT NOT NULL,
   details                     TEXT,
   warehoused_at               TEXT NOT NULL,
   updated_at                  TEXT NOT NULL,
-  source_id                   BIGINT NOT NULL,
+  source_id                   INTEGER NOT NULL,
   FOREIGN KEY(source_id) REFERENCES sources(id)
 );
 
 CREATE TABLE IF NOT EXISTS accounts (
-  id                          BIGINT PRIMARY KEY,
+  id                          INTEGER PRIMARY KEY,
   name                        TEXT NOT NULL,
   details                     TEXT,
-  currency_id                 BIGINT NOT NULL,
-  counterparty_id             BIGINT NOT NULL,
+  currency_id                 INTEGER NOT NULL,
+  counterparty_id             INTEGER NOT NULL,
   warehoused_at               TEXT NOT NULL,
   updated_at                  TEXT NOT NULL,
-  source_id                   BIGINT NOT NULL,
+  source_id                   INTEGER NOT NULL,
   FOREIGN KEY(currency_id) REFERENCES currencies(id),
   FOREIGN KEY(counterparty_id) REFERENCES counterparties(id),
   FOREIGN KEY(source_id) REFERENCES sources(id)
 );
 
 CREATE TABLE IF NOT EXISTS transactions (
-  id                          BIGINT PRIMARY KEY,
+  id                          INTEGER PRIMARY KEY,
   title                       TEXT NOT NULL,
   details                     TEXT,
-  withdrawn_account_id        BIGINT NOT NULL,
-  deposited_account_id        BIGINT NOT NULL,
+  withdrawn_account_id        INTEGER NOT NULL,
+  deposited_account_id        INTEGER NOT NULL,
   payment_amount              REAL NOT NULL,
-  payment_currency_id         BIGINT NOT NULL,
+  payment_currency_id         INTEGER NOT NULL,
   hold_amount                 REAL NOT NULL,
   transaction_amount          REAL,
   transactioned_at            TEXT NOT NULL,
   accounted_at                TEXT,
   warehoused_at               TEXT NOT NULL,
   updated_at                  TEXT NOT NULL,
-  source_id                   BIGINT NOT NULL,
+  source_id                   INTEGER NOT NULL,
   FOREIGN KEY(withdrawn_account_id) REFERENCES accounts(id),
   FOREIGN KEY(deposited_account_id) REFERENCES accounts(id),
   FOREIGN KEY(payment_currency_id) REFERENCES currencies(id),
